fix: avoid sending unwanted GraphQL booking query when opening SmartFAQ
